Refactored configuration reloading behavior into a parent class for external content path validator
